pyproject.toml
[project]
name = "fsp-2025-fianl-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "httpx>=0.27.0",
    "numpy>=2.0.0",
    "openai>=1.0.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "ruff>=0.14.6",
    "scikit-learn>=1.5.0",
    "sentence-transformers>=2.7.0",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [
  ".venv",
]
format.quote-style = "single"
lint.exclude = [
  "PLR0913",
]
lint.select = [
  "F", # Pyflakes
  # "E/W", # pycodestyle
  "C90", # mccabe
  # "I", # isort
  # "N", # pep8-naming
  # "D", # pydocstyle
  "UP", # pyupgrade
  "YTT", # flake8-2020
  # "ANN", # flake8-annotations
  # "ASYNC", # flake8-async
  # "TRIO", # flake8-trio
  # "S", # flake8-bandit
  # "BLE", # flake8-blind-except
  # "FBT", # flake8-boolean-trap
  "B", # flake8-bugbear
  # "A", # flake8-builtins
  # "COM", # flake8-commas
  # "CPY", # flake8-copyright
  "C4", # flake8-comprehensions
  # "DTZ", # flake8-datetimez
  "T10", # flake8-debugger
  # "DJ", # flake8-django
  # "EM", # flake8-errmsg
  # "EXE", # flake8-executable
  # "FA", # flake8-future-annotations
  "ISC", # flake8-implicit-str-concat
  # "ICN", # flake8-import-conventions
#  "G", # flake8-logging-format
  "INP", # flake8-no-pep420
  # "PIE", # flake8-pie
  "T20", # flake8-print
  # "PYI", # flake8-pyi
  # "PT", # flake8-pytest-style
  # "Q", # flake8-quotes
  "RSE", # flake8-raise
  "RET", # flake8-return
  "SLF", # flake8-self
  # "SLOT", # flake8-slots
  "SIM", # flake8-simplify
  "TID", # flake8-tidy-imports
  # "TCH", # flake8-type-checking
  # "INT", # flake8-gettext
  # "ARG", # flake8-unused-arguments
  "PTH", # flake8-use-pathlib
  # "TD", # flake8-todos
  # "FIX", # flake8-fixme
  "ERA", # eradicate
  # "PD", # pandas-vet
  # "PGH", # pygrep-hooks
  "PL", # Pylint
  # "TRY", # tryceratops
  # "FLY", # flynt
  # "NPY", # NumPy-specific rules
  # "AIR", # Airflow
  # "PERF", # Perflint
  # "FURB", # refurb
  # "LOG", # flake8-logging
  "RUF" # Ruff-specific rules
]
lint.flake8-tidy-imports.ban-relative-imports = "all"
lint.flake8-bugbear.extend-immutable-calls = [
  "fastapi.Depends",
  "fastapi.params.Depends",
  "fastapi.Query",
  "fastapi.params.Query"
]
lint.mccabe.max-complexity = 20
